requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
art==5.2
async-timeout==3.0.1
attrs==21.2.0
brotlipy==0.7.0
cachetools==4.2.2
certifi==2020.12.5
conda==4.9.2
conda-build==3.21.4
cycler==0.10.0
dataclasses==0.6
decord==0.5.2
dnspython==2.1.0
efficientnet-pytorch==0.7.1
fsspec==2021.4.0
future==0.18.2
google-auth==1.30.0
google-auth-oauthlib==0.4.4
grad-cam==1.2.9
grpcio==1.37.1
jedi==0.17.0
joblib==1.0.1
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.2
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict==5.1.0
natsort==7.1.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.2.52
packaging==20.9
pandas==1.2.4
parso==0.7.0
pkginfo==1.7.0
protobuf==3.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycm==3.1
pycosat==0.6.3
pyDeprecate==0.3.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-etcd==0.4.5
pytorch-lightning==1.3.0
PyYAML==5.3.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel-yaml==0.15.87
scikit-learn==0.24.2
scipy==1.6.3
sklearn==0.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.1.0
torch==1.8.0
torch-summary==1.4.4
torchelastic==0.2.1
torchmetrics==0.3.2
torchtext==0.9.0
torchvision==0.9.0
ttach==0.0.3
Werkzeug==1.0.1
yarl==1.6.3
torch_summary==1.4.4
efficientnet_pytorch==0.7.1
tqdm==4.51.0
pytorch_lightning==1.3.0
numpy==1.21.0
Pillow==8.2.0
torchsummary==1.5.1
scikit-image
